fix: Fix bugs in ListHead merge operation
This patch is abit messy, the former implementation includes lots of
ownership errors. Also lack of concern of the header of a list, we
try to fix it this time including the following changes
* Check whether the list is empty before performing list_del_entry()
* Refactor list_splice_tail() so the borrow rules won't be violated
* Change the implementation of merge()

Currently the implementation of merge() isn't elegant enough, it still
deal with two lists with special concern for their list header.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>