list: Implementation of list_sort()
Finish the version 1 of list_sort(), which utilize the vector
datastructure instead of pointer to pointer to implement the
linux kernel's list_sort().

In this version of implementation, we should be careful that
each list sends into any method should be the list header
instead of the list itself, also every lists stored in
"pending" vector is started with a list header.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>