Implementation of UniqueArc<T>
Since "Arc<T>" doesn't allow modification after construction and before it's shared,
there's no proper way to perform initialisation of it.

Introduce "UniqueArc<T>" which is essentially an "Arc<T>" with refcount value of 1,
so it's writeable, and can be safely transformed into an "Arc<T>" once the initialisation
is done.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>