sync: arc: Implement write() and assume_init() for UniqueArc
Implementation of write() and assume_init() method for UniqueArc,
write() is for proper initialization for UniqueArc, assume_init()
can check whether the instance is properly initialized or not,
UB should be triggered immediately if it isn't, however, for the
current implementation, the compiler won't complain. Fix it in
the future iteration.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>